requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
httpx==0.27.0
openai
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0